FIX TreeDropdownField cached items
When searching a tree dropdown field, and that dropdown field is looking at
the hierarchy of a class type that has already been iterated on that request,
the cached data present in Hierarchy will mean the search results do not
correctly mark items. Simplest solution is to clear the hierarchy cache.